Refactor ThanosSerializerContext, WarArena, and WarSnake; change capacity type to int for better compatibility, improve memory allocation logic, and enhance snake initialization process

+++ Thanos/SourceGen/ThanosSerializerContext.cs
@@ -111,7 +111,7 @@ public readonly struct Board(uint height, uint width, Coordinate[] food, Coordin
     [JsonPropertyName("snakes")] public readonly Snake[] Snakes = snakes;
 
     [JsonIgnore] public readonly uint Area = height * width;
-    [JsonIgnore] public readonly uint Capacity = Math.Min(BitOperations.RoundUpToPowerOf2(height * width), Constants.MaxBodyLength);
+    [JsonIgnore] public readonly int Capacity = (int)Math.Min(BitOperations.RoundUpToPowerOf2(height * width), Constants.MaxBodyLength);
 }
 
 [method: JsonConstructor]
